Move CompareMainTitle to new file
