+++ src/components/merge_results/fields/CompareFields.tsx
